Extract ipc json tests into their own test project.
